backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery==5.3.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
websockets==12.0
aiohttp==3.9.1

# AI & LLM
langgraph==0.0.43
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.6
openai==1.10.0
anthropic==0.18.0
google-generativeai==0.3.2

# Data Analysis
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
ta-lib==0.4.28

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
